Sprint 8 (#591 F3): extract SettingsActionHandler from Settings
Moves the settings action/maintenance handler cluster out of the Settings
god-class into a new SettingsActionHandler (handle_settings_submission,
handle_clear_qr_cache, handle_migration_execution, handle_obsolete_shortcode_
cleanup, handle_url_shortener_cleanup, handle_public_access_disabler,
handle_submission_link_audit, handle_cache_actions, ajax_preview_date_format).
Settings (1078 → 545) composes the handler and keeps same-named thin delegators,
so every add_action hook registration and existing test call is untouched. The
handler receives the already-built SettingsSaveHandler. Nonce/cap checks,
redirects, option writes and transients are unchanged.

Handler carries @covers + a setUp preload (verified 240/385; combined Settings
coverage rises vs the pre-split baseline). Behavior-preserving — Settings suites
green (31 tests); PHPStan L8 + WPCS clean.
